+++ src/game/tests.rs
@@ -443,6 +443,7 @@ mod vintage_cube_legolas_reflexes;
 mod vintage_cube_leyline_binding;
 mod vintage_cube_library;
 mod vintage_cube_lion_sash;
+mod vintage_cube_lions_eye_diamond;
 mod vintage_cube_loot_pathfinder;
 mod vintage_cube_loran;
 mod vintage_cube_lorien;